requirements.txt
telethon>=1.34.0
python-dotenv>=1.0.0
langchain-mistralai>=0.1.0
langchain-core>=0.2.0
python-telegram-bot==21.5
httpx>=0.24.0